+++ flask_template/db/cli.py
@@ -1,8 +1,6 @@
 """CLI functions for the db module."""
 
-from typing import cast
 from flask import Flask, Blueprint, current_app
-from flask.cli import with_appcontext, AppGroup
 import click
 
 from ..util.logging import get_logger
@@ -14,13 +12,12 @@ from . import models  # noqa
 
 
 DB_CLI_BLP = Blueprint("db_cli", __name__, cli_group=None)
-DB_CLI = cast(AppGroup, DB_CLI_BLP.cli)  # expose as attribute for autodoc generation
+DB_CLI = DB_CLI_BLP.cli  # expose as attribute for autodoc generation
 
 DB_COMMAND_LOGGER = "db"
 
 
 @DB_CLI.command("create-db")
-@with_appcontext
 def create_db():
     """Create all db tables."""
     create_db_function(current_app)
@@ -33,7 +30,6 @@ def create_db_function(app: Flask):
 
 
 @DB_CLI.command("drop-db")
-@with_appcontext
 def drop_db():
     """Drop all db tables."""
     drop_db_function(current_app)