+++ flask_template/db/db.py
@@ -1,11 +1,12 @@
 """Module to hold DB constant to avoid circular imports."""
 
 from typing import Type, cast
-from flask_sqlalchemy.model import DefaultMeta, Model
-from sqlalchemy.schema import MetaData
-from flask_sqlalchemy import SQLAlchemy
+
 from flask_migrate import Migrate
-from sqlalchemy.orm.decl_api import registry
+from flask_sqlalchemy import SQLAlchemy
+from flask_sqlalchemy.model import Model
+from sqlalchemy.orm import DeclarativeBase, registry
+from sqlalchemy.schema import MetaData
 
 
 DB: SQLAlchemy = SQLAlchemy(
@@ -22,13 +23,13 @@ DB: SQLAlchemy = SQLAlchemy(
 
 
 # Model constant to be importable directly
-MODEL = cast(Type[Model], DB.Model)
-if type(MODEL) is not DefaultMeta or not issubclass(MODEL, Model):
+MODEL = cast(Type[DeclarativeBase], DB.Model)
+if not issubclass(MODEL, Model):
     raise Warning(
         f"Please update the type cast of db.MODEL to reflect the current type {type(MODEL)}."
     )
 
-# only for sqlalchemy 1.4!
-REGISTRY = cast(registry, MODEL.registry)
+# remove this if you do not use the `REGISTRY.mapped_as_dataclass` decorator
+REGISTRY: registry = MODEL.registry
 
 MIGRATE = Migrate()