+++ flask_template/db/models/example.py
@@ -1,27 +1,25 @@
-from dataclasses import dataclass, field
 from typing import Optional
-from sqlalchemy.sql.schema import Column
+
+from sqlalchemy.orm import Mapped, mapped_column
 from sqlalchemy.sql import sqltypes as sql
 
 from ..db import MODEL, REGISTRY
 
 
+# see <https://docs.sqlalchemy.org/en/20/orm/mapping_styles.html#declarative-mapping>
 class Example(MODEL):
     __tablename__ = "Example"
-    id: Column = Column(sql.Integer, primary_key=True)
-    name: Column = Column(sql.String(120))
+    id: Mapped[int] = mapped_column(primary_key=True)
+    name: Mapped[str] = mapped_column(sql.String(120))
 
 
-# only for sqlalchemy 1.4 (see <https://docs.sqlalchemy.org/en/14/orm/mapping_styles.html#example-two-dataclasses-with-declarative-table>)
-# This style probably has better type hint and autocompletion support
-@REGISTRY.mapped
-@dataclass
+# see <https://docs.sqlalchemy.org/en/20/orm/dataclasses.html#integration-with-dataclasses-and-attrs>
+# use this style if you have 1.4 dataclass models to migrate
+@REGISTRY.mapped_as_dataclass
 class TestDataclass:
     __tablename__ = "TestDataclass"
 
-    __sa_dataclass_metadata_key__ = "sa"
-
-    id: int = field(init=False, metadata={"sa": Column(sql.Integer, primary_key=True)})
-    name: Optional[str] = field(default=None, metadata={"sa": Column(sql.String(50))})
-    fullname: str = field(default="", metadata={"sa": Column(sql.String(50))})
-    nickname: str = field(default="", metadata={"sa": Column(sql.String(12))})
+    id: Mapped[int] = mapped_column(primary_key=True)
+    name: Mapped[Optional[str]] = mapped_column(sql.String(50), default=None)
+    fullname: Mapped[str] = mapped_column(sql.String(50), default="")
+    nickname: Mapped[str] = mapped_column(sql.String(12), default="")